Create README.md
Hi
I'm David Alexander Hutley and I am 80 year old and therefore I am not writing this as a means to get a job but rather to explain why I joined this course and what I hope to achieve from it.
I started working in IT when I was 23 .   At that time I joined IBM New Zealand as an Analyst Programmer.   It that time in Auckland we had no commercial  computers installed although some businesses in Auckland had them on order.